Lombok added, refactored.

+++ src/main/java/by/edik/car_api/model/Picture.java
@@ -1,7 +1,7 @@
 package by.edik.car_api.model;
 
 public class Picture {
-    private int pictureId;
-    private int adId;
+    private long pictureId;
+    private long adId;
     private String pictureReference;
 }
