+++ src/main/java/by/edik/car_api/model/UserPhone.java
@@ -1,7 +1,7 @@
 package by.edik.car_api.model;
 
 public class UserPhone {
-    private int phoneNumberId;
-    private int userId;
+    private long phoneNumberId;
+    private long userId;
     private String phoneNumber;
 }
